backend/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch
transformers
fastapi
uvicorn
gunicorn
pydantic
scikit-learn
pandas
numpy
joblib
slowapi
bleach
python-dotenv
requests
sqlalchemy
psycopg2-binary
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
redis
cachetools
python-whois
